feat (VIZ) basic elements now use a Point object for positionning
