+++ viz/d2v-viz-common/src/main/kotlin/io/data2viz/viz/LineNode.kt
@@ -1,18 +1,20 @@
 package io.data2viz.viz
 
-import io.data2viz.color.ColorOrGradient
+import io.data2viz.color.*
+import io.data2viz.geom.*
 
 
 class LineNode : Node(),
-        HasStroke,
-        HasTransform {
+    HasStroke,
+    HasTransform,
+    HasStartAndEnd {
 
     override var transform: Transform? = null
 
-    var x1: Double = .0
-    var y1: Double = .0
-    var x2: Double = .0
-    var y2: Double = .0
+    override var x1: Double = .0
+    override var y1: Double = .0
+    override var x2: Double = .0
+    override var y2: Double = .0
 
     override var stroke: ColorOrGradient?
         get() = style.stroke
